refactor: remove unneccesary processes

+++ crates/analyzer/src/constants.rs
@@ -20,7 +20,7 @@ pub const MAX_NUMPAD_DIGIT: u8 = 9;
 pub const DEFAULT_FKEYS_MAX: u8 = 12;
 
 /// unit conversion u -> mm
-pub const U2MM: f64 = 19.05; // u -> mm (f64 for Fitts calculation)
+pub const U2MM: f32 = 19.05; // u -> mm (f64 for Fitts calculation)
 pub const U2CELL: usize = 4; // u -> cell
 pub const U2PX: f32 = 60.0; // u -> px (for visualization)
 
@@ -38,8 +38,8 @@ pub const FONT_SIZE: f32 = 16.0; // font size [px]
 /// - col: cell unit
 #[inline]
 pub fn cell_to_coordinate(row: usize, col: usize) -> (f32, f32) {
-    let x = (col as f32 / U2CELL as f32) * U2MM as f32;
-    let y = row as f32 * U2MM as f32;
+    let x = (col as f32 / U2CELL as f32) * U2MM;
+    let y = row as f32 * U2MM;
     (x, y)
 }
 
@@ -50,8 +50,8 @@ pub fn cell_to_coordinate(row: usize, col: usize) -> (f32, f32) {
 #[inline]
 pub fn cell_to_key_center(row: usize, col: usize, width: f32) -> (f32, f32) {
     let (mut x, mut y) = cell_to_coordinate(row, col);
-    x += width / 2.0 * U2MM as f32;
-    y += 0.5 * U2MM as f32; // キーの中心位置（0.5u offset for center）
+    x += width / 2.0 * U2MM;
+    y += 0.5 * U2MM; // キーの中心位置（0.5u offset for center）
     (x, y)
 }
